+++ package.go
@@ -1,4 +1,4 @@
-/* Package qual implements tests and helpers for writing highly readable tests.
+/* Package qual implements tests and helpers for writing highly readable code.
 
  */
 package qual