+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Changed
 
+- LineLength only shows which line to trim and by how much
 - Assert() removed, resulted in Not so readable output
 
 ## [0.2.1] 2018-12-15